Application Search Updates
1. Relocate Application Search into the application folder
2. Update the Application Search response to use the PagedResponseDto base class
3. Update logging
4. Add error handling.
