+++ backend/cats/src/app/cats.module.ts
@@ -51,8 +51,8 @@ import { Payment } from './entities/payment.entity';
 import { PaymentMethod } from './entities/paymentMethod.entity';
 import { Organization } from './entities/organization.entity';
 import { ParticipantRole } from './entities/participantRole.entity';
-import { ApplicationSearchResolver } from './resolvers/applicationSearch.resolver';
-import { ApplicationSearchService } from './services/applicationSearch.service';
+import { ApplicationSearchResolver } from './resolvers/application/applicationSearch.resolver';
+import { ApplicationSearchService } from './services/application/applicationSearch.service';
 import { AppParticipantService } from './services/application/appParticipants.service';
 import { AppParticipantResolver } from './resolvers/application/appParticipant.resolver';
 import { GenericResponseProvider } from './dto/response/genericResponseProvider';
