+++ backend/cats/src/app/dto/response/applicationSearchResponse.ts
@@ -1,17 +1,17 @@
-import { Field, ObjectType, Int } from '@nestjs/graphql';
+import { Field, ObjectType } from '@nestjs/graphql';
 import { ApplicationResultDto } from '../applicationResultDto';
+import { PagedResponseDto } from './response.dto';
+
+export class ApplicationSearchResult {
+  applications: ApplicationResultDto[];
+  count: number;
+  page: number;
+  pageSize: number;
+  error?: string;
+}
 
 @ObjectType()
-export class ApplicationSearchResponse {
+export class ApplicationSearchResponse extends PagedResponseDto {
   @Field(() => [ApplicationResultDto])
   applications: ApplicationResultDto[];
-
-  @Field(() => Int)
-  count: number;
-
-  @Field(() => Int)
-  page: number;
-
-  @Field(() => Int)
-  pageSize: number;
 }
